requirements.txt
openai
elevenlabs
streamlit
requests
python-dotenv